refactor(backup): extract export flow and save dialog into backup/export.rs
